enemies working
features for each enemy:

causeDamage: input is a number with the damage caused to the enemy, and the value inputted will be subtracted from the health of the enemy. if the health reaches to 0 or below, the enemy is dead.

flip: flip the enemy from facing the right to the left.

getHealth: get current health of the enemy.

update: call this function in the scene's update function to shoot pellets at a certain fireRate.